refactor: split codegen support inventories
